+++ src/Symfony/Component/Finder/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+7.4
+---
+
+ * Extend `name()` and `notName()` to allow overwriting existing names/notNames filters.
+ * Added `resetName()` and `resetNotName()` to reset names/notNames filters
+
 6.4
 ---